lld: Always enable build-id and use 20-byte hashes
Always ensure that build-id is used, and ensure that is is 20bytes in length. If the user does not want this behavior they can use `--build-id=none` explicitly.

Note that though this patch uses "sha1" LLVM actually uses the BLAKE3 algorithm and truncates the output after the first 20 bytes.

It would be nice if llvm supported configuring this at build time like binutils does.

Origin: vendor
Forwarded: not-needed
Last-Update: 2025-03-29