Use mimalloc for clang/lld/llvm
In testing this resulted in approximately a 7% improvement in build times for a few simple tests.

Downstream-only because upstream prefers to support this by passing in a static linking flag to use libmimalloc.a, but as a distro we prefer to use the shared library.

Origin: vendor
Forwarded: not-needed
Last-Update: 2025-03-29